[#21]/reafactor: schoolpages
