+++ src/apis/school-pages/controllers/school-pages.controller.spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
+import { SchoolPageType } from 'src/apis/school-pages/constants/school-page.enum';
 import { CreateSchoolPageNewsRequestBodyDto } from 'src/apis/school-pages/dto/create-school-page-news-request-body.dto';
 import { CreateSchoolPageRequestBodyDto } from 'src/apis/school-pages/dto/create-school-page-request-body.dto';
 import { FindAllSchoolPageNewsRequestQueryDto } from 'src/apis/school-pages/dto/find-all-school-page-news-request-query.dto';
 import { FindAllSchoolPageRequestQueryDto } from 'src/apis/school-pages/dto/find-all-school-page-request-query.dto';
 import { PartialUpdateSchoolPageNewsRequestBodyDto } from 'src/apis/school-pages/dto/partial-update-school-page-news-request-body.dto';
-import { SchoolType } from 'src/constants/school/school.enum';
 import { StudentEntity } from 'src/entities/student.entity';
 import { MockSchoolPagesService } from 'test/mock/mock.service';
+import { SchoolPagesService } from '../school-pages.service';
 import { SchoolPagesController } from './school-pages.controller';
-import { SchoolPagesService } from './school-pages.service';
 
 describe(SchoolPagesController.name, () => {
   let controller: SchoolPagesController;
@@ -77,12 +77,12 @@ describe(SchoolPagesController.name, () => {
 
     it('전체 조회 성공', async () => {
       student.id = 1;
-      findAllSchoolPageRequestQueryDto.type = SchoolType.Elementary;
+      findAllSchoolPageRequestQueryDto.type = SchoolPageType.Elementary;
 
       const schoolPages = [
         {
           id: 2,
-          type: SchoolType.Elementary,
+          type: SchoolPageType.Elementary,
         },
       ];
