+++ src/apis/school-pages/dto/create-school-page-news-request-body.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsString, Length } from 'class-validator';
-import { SchoolPageNewsEntity } from 'src/entities/school-news.entity';
+import { SchoolPageNewsEntity } from 'src/apis/school-pages/entities/school-news.entity';
 
 export class CreateSchoolPageNewsRequestBodyDto
   implements Pick<SchoolPageNewsEntity, 'title' | 'description'>
