+++ src/apis/school-pages/dto/create-school-page-request-body.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEnum, IsString, IsUrl, Length } from 'class-validator';
-import { SchoolType } from 'src/constants/school/school.enum';
-import { SchoolPageEntity } from 'src/entities/school-page.entity';
+import { SchoolPageType } from 'src/apis/school-pages/constants/school-page.enum';
+import { SchoolPageEntity } from 'src/apis/school-pages/entities/school-page.entity';
 
 export class CreateSchoolPageRequestBodyDto
   implements
@@ -21,10 +21,10 @@ export class CreateSchoolPageRequestBodyDto
 
   @ApiProperty({
     description: '학교 타입',
-    enum: SchoolType,
+    enum: SchoolPageType,
   })
-  @IsEnum(SchoolType)
-  type: SchoolType;
+  @IsEnum(SchoolPageType)
+  type: SchoolPageType;
 
   @ApiProperty({
     description: '주소',
