+++ src/apis/school-pages/school-pages.service.ts
@@ -11,10 +11,10 @@ import { CreateSchoolPageRequestBodyDto } from 'src/apis/school-pages/dto/create
 import { FindAllSchoolPageNewsRequestQueryDto } from 'src/apis/school-pages/dto/find-all-school-page-news-request-query.dto';
 import { FindAllSchoolPageRequestQueryDto } from 'src/apis/school-pages/dto/find-all-school-page-request-query.dto';
 import { PartialUpdateSchoolPageNewsRequestBodyDto } from 'src/apis/school-pages/dto/partial-update-school-page-news-request-body.dto';
-import { SchoolPageNewsEntity } from 'src/entities/school-news.entity';
-import { SchoolPageAdminLinkEntity } from 'src/entities/school-page-admin-link.entity';
-import { SchoolPageSubscribeLinkEntity } from 'src/entities/school-page-subscribe-link.entity';
-import { SchoolPageEntity } from 'src/entities/school-page.entity';
+import { SchoolPageNewsEntity } from 'src/apis/school-pages/entities/school-news.entity';
+import { SchoolPageAdminLinkEntity } from 'src/apis/school-pages/entities/school-page-admin-link.entity';
+import { SchoolPageSubscribeLinkEntity } from 'src/apis/school-pages/entities/school-page-subscribe-link.entity';
+import { SchoolPageEntity } from 'src/apis/school-pages/entities/school-page.entity';
 import { DataSource, FindOptionsWhere, Repository } from 'typeorm';
 
 @Injectable()
