+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppConfigModule } from 'src/config/app-config/app-config.module';
 import { TypeOrmConfigService } from 'src/config/external-config/typeorm-config.service';
 import { AuthModule } from './apis/auth/auth.module';
-import { SchoolPagesModule } from './apis/school-pages/school-pages.module';
+import { SchoolPagesModule } from './apis/school-pages/services/school-pages.module';
 import { StudentsModule } from './apis/students/students.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
